move enrichment router into a separate module add more comments to swagger UI related to #686
